+++ examples/client/src/types.ts
@@ -1,3 +1,0 @@
-export interface Extensions {
-  version: string;
-}