+++ examples/server/gql2ts.js
@@ -1,8 +1,0 @@
-/* eslint-disable @typescript-eslint/no-var-requires */
-const { gql2ts } = require("../.."); // graphqlade in your app
-
-gql2ts({
-  root: __dirname,
-  server: true,
-  stitching: true,
-});