+++ examples/server/src/resolvers/index.ts
@@ -3,7 +3,7 @@ import { MyContext } from "../MyContext";
 import { mutationResolvers } from "./Mutation";
 import { queryResolvers } from "./Query";
 import { reviewResolvers } from "./reviewResolvers";
-import { scalarResolvers } from "./scalars";
+import { scalarResolvers } from "./scalarResolvers";
 
 export const resolvers: Resolvers<MyContext>[] = [
   queryResolvers,